VASP/requirements.txt
torch==1.12.1
torchvision==0.13.1
pandas
scikit-learn
wandb
hydra-core
bottleneck